Use different OAV fixture in system tests
